Test cases for container methods
